@@ -1,32 +1,19 @@
-const unifi = require("node-unifi");
-const Bluebird = require("bluebird");
-const R = require("ramda");
-const retry = require("bluebird-retry");
+import Bluebird from "bluebird";
+import retry from "bluebird-retry";
+import unifi from "node-unifi";
+import { compose, filter, flatten } from "ramda";
 
-let Accessory, Service, Characteristic, UUIDGen;
+const SECOND = 1000;
+const DEFAULT_INTERVAL = 60 * SECOND;
+const DEFAULT_PORT = 443;
 
-module.exports = function (homebridge) {
-  console.log("homebridge API version: " + homebridge.version);
-
-  // Accessory must be created from PlatformAccessory Constructor
-  Accessory = homebridge.platformAccessory;
-
-  // Service and Characteristic are from hap-nodejs
-  Service = homebridge.hap.Service;
-  Characteristic = homebridge.hap.Characteristic;
-  UUIDGen = homebridge.hap.uuid;
-
-  // For platform plugin to be considered as dynamic platform plugin,
-  // registerPlatform(pluginName, platformName, constructor, dynamic), dynamic must be true
-  homebridge.registerPlatform("homebridge-unifi-guest-wifi-platform", "UnifyGuestWifiPlatform", UnifyGuestWifiPlatform, true);
-};
+let Accessory;let Service;let Characteristic;let UUIDGen;
 
 // Platform constructor
 // config may be null
 // api may be null if launched from old homebridge version
 function UnifyGuestWifiPlatform(log, config, api) {
   log("UnifyGuestWifiPlatform Init");
-  var platform = this;
   this.log = log;
   this.config = config;
   this.accessories = {};
@@ -36,8 +23,8 @@ function UnifyGuestWifiPlatform(log, config, api) {
 
   this.unifiController = new unifi.Controller({
     host: this.controllerConfig.address,
-    port: this.controllerConfig.port || 8443,
-    sslverify: false
+    port: this.controllerConfig.port || DEFAULT_PORT,
+    sslverify: this.controllerConfig.sslVerify || false
   });
 
   // Bluebird.promisifyAll(this.unifiController)
@@ -55,7 +42,7 @@ function UnifyGuestWifiPlatform(log, config, api) {
         return this.unifiController.login(this.controllerConfig.username, this.controllerConfig.password).then(() => {
           this.log("Logged into Unifi Controller");
           return this.loadGuestWifi().then(() => {
-            const interval = this.controllerConfig.updateInterval || 1000000;
+            const interval = this.controllerConfig.updateInterval * SECOND || DEFAULT_INTERVAL;
 
             this.log(`Setting up update interval: ${interval}`);
 
@@ -64,11 +51,11 @@ function UnifyGuestWifiPlatform(log, config, api) {
               this.loadGuestWifi();
             }, interval);
 
-            platform.log("DidFinishLaunching");
+            this.log("DidFinishLaunching");
           });
-        }).catch(e => {
-          this.log("error loading guest wifi", e);
-          throw e;
+        }).catch(error => {
+          this.log("error loading guest wifi", error);
+          throw error;
         });
       }, {
         interval: 6000,
@@ -76,19 +63,20 @@ function UnifyGuestWifiPlatform(log, config, api) {
         throw_original: true
       }).then(() => {
         this.ready = true;
-      });
+        return true;
+      }).catch(this.log);
     });
   }
 }
 
 UnifyGuestWifiPlatform.prototype.loadGuestWifi = async function () {
-  this.log("loading guest wifies");
+  // this.log("loading guest wifies")
 
   let wlans = await this.unifiController.getWLanSettings();
 
-  wlans = R.compose(R.filter(wlan => wlan.is_guest), R.flatten)(wlans);
+  wlans = compose(filter(wlan => wlan.is_guest), flatten)(wlans);
 
-  this.log("loaded wlan settings"); // , wlans)
+  // this.log("loaded wlan settings") // , wlans)
 
   if (wlans.length <= 0) {
     return;
@@ -100,7 +88,7 @@ UnifyGuestWifiPlatform.prototype.loadGuestWifi = async function () {
 };
 
 UnifyGuestWifiPlatform.prototype.addGuestWifiAccessories = async function (wlans) {
-  this.log(`adding ${wlans.length} guest wifi`);
+  // this.log(`adding ${wlans.length} guest wifi`)
 
   await Bluebird.all(wlans.map(wlan => this.addGuestWifiAccessory(wlan)));
 };
@@ -110,19 +98,19 @@ UnifyGuestWifiPlatform.prototype.generateAccessoryName = wlan => wlan.name;
 UnifyGuestWifiPlatform.prototype.generateAccessoryId = wlan => UUIDGen.generate(wlan._id);
 
 UnifyGuestWifiPlatform.prototype.setupAccessory = function (accessory, configure) {
-  const context = accessory.context;
+  const { context, UUID } = accessory;
 
   this.log(`Setting up accessory for Guest Wifi: ${this.generateAccessoryName(context.wlan)}`);
 
   accessory.on("identify", (paired, callback) => {
-    this.log(newAccessory.displayName, "Identify!!!");
+    // this.log(newAccessory.displayName, "Identify!!!")
     callback();
   });
   // Plugin can save context on accessory to help restore accessory in configureAccessory()
   // newAccessory.context.something = "Something"
 
   // Make sure you provided a name for service, otherwise it may not visible in some HomeKit apps
-  this.accessories[accessory.UUID] = accessory;
+  this.accessories[UUID] = accessory;
 
   this.log(`getting service switch ${accessory.getService(Service.Switch)}, ${configure}`);
 
@@ -132,32 +120,32 @@ UnifyGuestWifiPlatform.prototype.setupAccessory = function (accessory, configure
 
   accessory.getService(Service.Switch).getCharacteristic(Characteristic.On).on("set", async (value, callback) => {
     if (!this.ready) {
-      callback(new Error("unifi controller is not ready yet."));
+      callback(new Error("Set Failed: Unifi Controller is not ready yet."));
     }
 
-    const { wlan } = accessory.context;
+    const { context: { wlan }, displayName } = accessory;
 
-    const wlanName = this.generateAccessoryName(context.wlan);
+    const wlanName = this.generateAccessoryName(wlan);
 
-    this.log(`${accessory.displayName} requested ${value} id ${wlan._id.trim()}`);
+    this.log(`${displayName} requested ${value} id ${wlan._id.trim()}`);
 
     await this.unifiController.disableWLan(String(wlan._id), !value);
 
     wlan.enabled = value;
 
-    this.log(accessory.displayName, `Set Guest Wifi -> ${wlanName}, ${value}`);
+    this.log(displayName, `Set Guest Wifi -> ${wlanName}, ${value}`);
 
     callback();
   }).on("get", callback => {
     if (!this.ready) {
-      this.log("unifi controller is not ready yet.");
-      callback(new Error("unifi controller is not ready yet."));
+      // this.log("unifi controller is not ready yet.")
+      callback(new Error("Get Failed: unifi controller is not ready yet."));
       return;
     }
 
     const { wlan } = accessory.context;
 
-    this.log(accessory.displayName, `Get Guest Wifi -> ${this.generateAccessoryName(wlan)}, ` + wlan.enabled);
+    this.log(accessory.displayName, `Get Guest Wifi -> ${this.generateAccessoryName(wlan)}, ${wlan.enabled}`);
 
     callback(null, wlan.enabled);
   });
@@ -178,7 +166,7 @@ UnifyGuestWifiPlatform.prototype.addGuestWifiAccessory = async function (wlan) {
 
   this.log(`Guest Wifi: ${name}, does not exists, adding new accessory`);
 
-  var newAccessory = new Accessory(name, uuid);
+  const newAccessory = new Accessory(name, uuid);
 
   newAccessory.context = { wlan };
 
@@ -197,7 +185,7 @@ UnifyGuestWifiPlatform.prototype.registerAccessory = function (accessory) {
 // Developer can configure accessory at here (like setup event handler).
 // Update current value.
 UnifyGuestWifiPlatform.prototype.configureAccessory = function (accessory) {
-  this.log(accessory.displayName, "Configure Accessory", accessory);
+  this.log(accessory.displayName, "Configure Accessory"); // , accessory)
 
   if (this.accessories[accessory.UUID]) {
     return;
@@ -206,17 +194,18 @@ UnifyGuestWifiPlatform.prototype.configureAccessory = function (accessory) {
   // Set the accessory to reachable if plugin can currently process the accessory,
   // otherwise set to false and update the reachability later by invoking
   // accessory.updateReachability()
-  accessory.reachable = true;
+  accessory.reachable = true; // eslint-disable-line no-param-reassign
 
   this.setupAccessory(accessory, true);
 };
 
 UnifyGuestWifiPlatform.prototype.updateAccessoriesReachability = function () {
-  this.log("Update Reachability");
-  for (var index in this.accessories) {
-    var accessory = this.accessories[index];
-    accessory.updateReachability(false);
-  }
+  // this.log("Update Reachability")
+  // for (const index in this.accessories) {
+  //   const accessory = this.accessories[index]
+  //   accessory.updateReachability(false)
+  // }
+  Object.values(this.accessories).forEach(accessory => accessory.updateReachability(false));
 };
 
 // Sample function to show how developer can remove accessory dynamically from outside event
@@ -226,3 +215,19 @@ UnifyGuestWifiPlatform.prototype.removeAccessory = function () {
 
   this.accessories = [];
 };
+
+export default function unifiGuestWlan(homebridge) {
+  // console.log(`homebridge API version: ${  homebridge.version}`)
+
+  // Accessory must be created from PlatformAccessory Constructor
+  Accessory = homebridge.platformAccessory;
+
+  // Service and Characteristic are from hap-nodejs
+  Service = homebridge.hap.Service;
+  Characteristic = homebridge.hap.Characteristic;
+  UUIDGen = homebridge.hap.uuid;
+
+  // For platform plugin to be considered as dynamic platform plugin,
+  // registerPlatform(pluginName, platformName, constructor, dynamic), dynamic must be true
+  homebridge.registerPlatform("homebridge-unifi-guest-wifi-platform", "UnifyGuestWifiPlatform", UnifyGuestWifiPlatform, true);
+}
